lang_graph_poc_stock_tech_analysis/requirements.txt
langgraph
langchain
langchain-community
chromadb
sentence-transformers
yfinance
pandas
numpy
ta
matplotlib
vaderSentiment
ragas
google-generativeai
litellm
